refactor: utils 패키지 클래스들 domain으로 이전
