+++ src/main/java/domain/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package utils;
+package domain;
 
 public class RandomNumberGenerator {
