+++ src/main/java/domain/RepeatCount.java
@@ -1,4 +1,4 @@
-package utils;
+package domain;
 
 public class RepeatCount {
