+++ src/main/java/domain/StringParser.java
@@ -1,4 +1,4 @@
-package utils;
+package domain;
 
 import java.util.Arrays;
 import java.util.List;
