+++ src/test/java/domain/NamesTest.java
@@ -1,4 +1,4 @@
-package utils;
+package domain;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
