+++ src/test/java/domain/RepeatCountTest.java
@@ -1,4 +1,4 @@
-package utils;
+package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
